Authorization Deleting a User's Task

+++ src/tasks/tasks.service.ts
@@ -23,8 +23,8 @@ export class TasksService {
     throw new NotFoundException(`Task with ID "${id}" not found`);
   }
 
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
